GEODE-9925: Add LICENSE and NOTICE for binary (#18)
The kafka connector binary zip distribution includes apache geode and all it's
dependencies, some of which are not apache licensed. Creating a LICENSE and
NOTICE for the binary distribution, which are copied from Apache Geodes LICENSE
and NOTICE. Changing the build to include these files in the doc directory of
the binary distribution.